requirement.txt
python-dotenv
pandas
numpy
pymongo
certifi
pymongo[srv]
scikit-learn
#-e .